completed the task
completed the task2

+++ no_one_academy_task2/lib/model/product.dart
@@ -53,7 +53,7 @@ class Rating {
 
   factory Rating.fromJson(Map<String, dynamic> json) {
     return Rating(
-      rate: json['rate'].toDouble() as double,
+      rate: json['rate'],
       count: json['count'] as int,
     );
   }
